dom/grid/test/chrome/test_grid_repeats.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
<style>
body {
	margin: 40px;
}
.wrapper {
	display: grid;
	width: 600px;
	grid-gap: 10px;
	grid-template-columns: repeat(2, 100px) repeat(auto-fill, 50px);
	background-color: #f00;
}
.box {
	background-color: #444;
	color: #fff;
}

</style>

<script>
'use strict';

SimpleTest.waitForExplicitFinish();

function runTests() {
	var wrapper = document.getElementById("wrapper");
	var grid = wrapper.getGridFragments()[0];
	
	// test state of tracks
	is(grid.cols.tracks[1].state, "static", "Grid column track 1 is marked as static.");
	is(grid.cols.tracks[2].state, "repeat", "Grid column track 2 is marked as repeat.");
	
	SimpleTest.finish();
}
</script>
</head>
<body onLoad="runTests();">

	<div id="wrapper" class="wrapper">
		<div id="boxA" class="box a">A</div>
		<div class="box b">B</div>
		<div class="box c">C</div>
		<div class="box d">D</div>
		<div class="box e">E</div>
	</div>
	
</body>
</html>
